src/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
pytest==8.0.0
pylint==3.0.3
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0
pydantic==1.10.13
